Updated unit tests

+++ Tests/SwiftOSCIONIOTests/TCP/Client and Server/OSCTCPServer Tests.swift
@@ -19,8 +19,8 @@ struct OSCTCPServer_Tests {
         let server = OSCTCPServer(port: nil, timeTagMode: .ignore)
 
         final actor Receiver {
-            var messages: [(message: OSCMessage, host: String, port: Int)] = []
-            func received(_ message: OSCMessage, host: String, port: Int) {
+            var messages: [(message: OSCMessage, host: String, port: UInt16)] = []
+            func received(_ message: OSCMessage, host: String, port: UInt16) {
                 messages.append((message, host, port))
             }
         }
